Dutch survival data

This data frame contains information about all Dutch who died above age 92 years between 1986 and 2015. Observations are doubly truncated and such bounds are calculated based on the range of plausible values for these variables. There are 226 records that are interval-censored and interval-truncated for which bdate, ddate and ndays is missing (NA).

Usage

dutch

Format

A data frame with 305143 rows and 11 variables:

ndays

survival time (in days)

bdate

the smallest plausible birth date given information about month of birth and death and survival (Date)

bmonth

month of birth

byear

year of birth

ddate

the largest plausible death date given information about month of birth and death and survival (Date)

dmonth

month of death

dyear

year of death

ltrunc

minimum age (in days); the maximum of either 92 years or the number of days reached in 1986

rtrunc

maximum age (in days) an individual could have reached by the end of 2015

gender

factor indicating gender of individual, either female or male

valid

quality flag; A for individuals born in the Netherlands, B for individuals born abroad who died in the Netherlands

Source

Statistics Netherlands (CBS). Accessed via the Supplemental material of Einmahl, Einmahl and de Haan (2019)

References

Einmahl, J.J., J.H.J. Einmahl and L. de Haan (2019). Limits to Human Life Span Through Extreme Value Theory, Journal of the American Statistical Association, 114(527), 1075-1080. https://doi.org/10.1080/01621459.2018.1537912