Sample observations from an interval truncated excess lifetime distribution

Sample observations from an interval truncated excess lifetime distribution

Usage

r_ditrunc_elife(
  n,
  scale,
  shape,
  lower,
  upper,
  family = c("exp", "gp", "gomp", "gompmake", "weibull", "extgp")
)

Arguments

n

sample size

scale

scale parameter(s)

shape

shape parameter(s)

lower

vector of lower bounds

upper

vector of upper bounds

family

string; choice of parametric family, either exponential (exp), Weibull (weibull), generalized Pareto (gp), Gompertz (gomp), Gompertz-Makeham (gompmake) or extended generalized Pareto (extgp).

Value

a vector of n observations

Examples

n <- 100L
# the lower and upper bound are either scalar,
# or else vectors of length n
r_dtrunc_elife(n = n, scale = 1, shape = -0.1,
               lower = pmax(0, runif(n, -0.5, 1)),
               upper = runif(n, 6, 10),
               family = "gp")
#>   [1] 0.58025009 2.59985618 1.24786952 0.72735563 0.92574431 1.73849416
#>   [7] 2.84178998 1.18222667 1.86012936 0.97052633 3.04722079 0.10564937
#>  [13] 0.59393785 1.69252041 0.32414608 1.05078377 0.96430925 0.84810836
#>  [19] 3.50172140 0.31389855 1.44191497 0.61242685 0.43111313 1.29195858
#>  [25] 0.86558521 0.12413623 1.00192036 1.24829229 1.74941939 3.16538912
#>  [31] 0.89496372 0.16207944 1.94883430 0.46192666 0.28241805 1.00576959
#>  [37] 1.58413652 0.53475080 0.61849133 0.12982633 0.95750412 0.64473116
#>  [43] 1.19281072 0.39354711 0.50456404 0.52242855 1.02790873 2.11439841
#>  [49] 1.31320329 0.80247167 0.33156666 1.85329470 3.04144057 0.68152476
#>  [55] 0.30284510 0.35818896 0.25471761 1.49836028 1.52264519 0.60478588
#>  [61] 1.16306984 2.82886081 1.53935393 0.20716444 1.15605402 4.28633989
#>  [67] 0.83817925 0.69511100 0.75207124 2.37857723 0.90045716 0.01069799
#>  [73] 1.30257105 2.37221523 1.50231844 1.79642989 1.14811245 1.37044963
#>  [79] 0.60100852 1.13586664 0.05160670 0.55058284 1.14345822 0.73845164
#>  [85] 1.82338410 1.33455938 3.36785663 0.07345253 2.80435514 2.98673757
#>  [91] 0.91169081 0.79072175 0.88138582 0.87276393 1.00144294 2.84326843
#>  [97] 0.86738734 0.35631843 1.09544771 1.35693429