Sample observations from an interval truncated excess lifetime distribution

Sample observations from an interval truncated excess lifetime distribution

Usage

r_dtrunc_elife(
  n,
  scale,
  rate,
  shape,
  lower,
  upper,
  family = c("exp", "gp", "gomp", "gompmake", "weibull", "extgp", "gppiece",
    "extweibull", "perks", "beard", "perksmake", "beardmake")
)

Arguments

n

sample size

scale

scale parameter(s)

rate

rate parameter(s)

shape

shape parameter(s)

lower

vector of lower bounds

upper

vector of upper bounds

family

string; choice of parametric family

Value

a vector of n observations

Examples

n <- 100L
# the lower and upper bound are either scalar,
# or else vectors of length n
r_dtrunc_elife(n = n, scale = 1, shape = -0.1,
               lower = pmax(0, runif(n, -0.5, 1)),
               upper = runif(n, 6, 10),
               family = "gp")
#>   [1] 1.45786773 0.63594744 6.35919638 0.57036658 0.60636925 1.25277276
#>   [7] 1.11713450 0.42443070 1.54756695 0.62978851 1.41410330 0.84346384
#>  [13] 0.93911835 1.80935024 0.94589386 1.13236809 0.58498821 0.49018195
#>  [19] 2.07202565 1.21573948 0.06622696 1.43297832 1.19289637 0.55857641
#>  [25] 0.13112440 0.61223167 0.51600594 1.31411918 0.43156669 1.03306446
#>  [31] 0.42707219 2.66551922 1.36231466 1.90474228 2.26792774 1.73592427
#>  [37] 1.20215771 1.70483255 1.13019956 0.54050902 1.97168922 1.39772100
#>  [43] 0.48337871 1.12778125 1.37120903 0.51905846 1.74161351 0.55933327
#>  [49] 1.52369337 0.61197903 0.47284826 1.58953155 0.45383271 0.97598781
#>  [55] 1.67138424 2.37315799 0.42913515 1.46461168 0.61192779 1.84292052
#>  [61] 1.60862868 2.13539655 0.38170322 1.04871880 0.77418532 0.86259746
#>  [67] 0.85372359 0.43824368 2.33262598 0.04178455 0.17990247 0.40503370
#>  [73] 1.50359853 1.14542350 0.55753929 0.14312248 0.07593681 0.65871700
#>  [79] 1.70593454 0.46476056 0.74276900 1.08455070 1.12886810 0.04843086
#>  [85] 2.75103949 0.95896819 3.24932055 0.33458625 0.76890192 0.36470188
#>  [91] 0.34179294 1.21286075 0.46038050 2.63537639 1.61526487 0.85707129
#>  [97] 0.29025528 0.72878781 0.52060492 0.59206529