Sample observations from a left-truncated right-censored excess lifetime distribution

Sample observations from a left-truncated right-censored excess lifetime distribution

Usage

r_ltrc_elife(
  n,
  scale,
  shape,
  lower,
  upper,
  family = c("exp", "gp", "gomp", "gompmake", "weibull", "extgp")
)

Arguments

n

sample size

scale

scale parameter

shape

shape parameter(s)

lower

vector of lower bounds

upper

vector of upper bounds above which data are right-truncated

family

string; choice of parametric family, either exponential (exp), Weibull (weibull), generalized Pareto (gp), Gompertz (gomp) or extended generalized Pareto (extgp).

Value

a list with n observations dat and a logical vector of the same length with TRUE for right-censored observations and FALSE otherwise.

Examples

n <- 100L
# the lower and upper bound are either scalar,
# or else vectors of length n
r_ltrc_elife(n = n, scale = 5, shape = -0.1,
               lower = pmax(0, runif(n, -0.5, 1)),
               upper = 5,
               family = "gp")
#> $dat
#>   [1] 0.83585031 1.70880263 1.42453756 0.23144122 5.00000000 2.30562718
#>   [7] 5.00000000 5.00000000 5.00000000 5.00000000 0.98456910 5.00000000
#>  [13] 4.23002522 3.79779690 5.00000000 3.22040734 2.36824187 4.92932685
#>  [19] 5.00000000 4.51399962 0.68632641 5.00000000 1.71437057 5.00000000
#>  [25] 1.31848157 4.16504406 5.00000000 5.00000000 5.00000000 5.00000000
#>  [31] 1.61895191 3.43188558 4.55637117 5.00000000 5.00000000 4.98305949
#>  [37] 5.00000000 2.17279614 3.41941684 0.33345183 0.91456334 3.90535302
#>  [43] 5.00000000 4.89610244 5.00000000 5.00000000 5.00000000 0.83252402
#>  [49] 5.00000000 2.94284346 5.00000000 5.00000000 3.47590448 4.78226746
#>  [55] 2.64897262 0.67223497 0.77181044 4.22477841 0.54207303 5.00000000
#>  [61] 5.00000000 5.00000000 5.00000000 5.00000000 0.06680135 1.91516450
#>  [67] 1.18540923 1.40385110 4.33768433 3.70974599 2.99813398 1.86198048
#>  [73] 5.00000000 5.00000000 4.67124875 1.23677220 4.40157209 5.00000000
#>  [79] 1.08246176 1.26011020 4.70711602 0.57334641 4.30333837 5.00000000
#>  [85] 1.91817546 1.73946531 4.15969778 3.26331840 5.00000000 4.55242661
#>  [91] 3.61251745 0.30208529 4.13543701 1.10502980 1.13631242 2.60020902
#>  [97] 1.17203675 5.00000000 4.13022425 5.00000000
#> 
#> $rcens
#>   [1] FALSE FALSE FALSE FALSE  TRUE FALSE  TRUE  TRUE  TRUE  TRUE FALSE  TRUE
#>  [13] FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE  TRUE FALSE  TRUE
#>  [25] FALSE FALSE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE  TRUE  TRUE FALSE
#>  [37]  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE  TRUE  TRUE  TRUE FALSE
#>  [49]  TRUE FALSE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73]  TRUE  TRUE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE  TRUE
#>  [85] FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE  TRUE FALSE  TRUE
#>