Sample observations from an interval truncated excess lifetime distribution

Sample observations from an interval truncated excess lifetime distribution

Usage

r_ditrunc_elife(
  n,
  rate,
  scale,
  shape,
  lower,
  upper,
  family = c("exp", "gp", "gomp", "gompmake", "weibull", "extgp", "gppiece",
    "extweibull", "perks", "beard", "perksmake", "beardmake")
)

Arguments

n

sample size

rate

rate parameter(s)

scale

scale parameter(s)

shape

shape parameter(s)

lower

vector of lower bounds

upper

vector of upper bounds

family

string; choice of parametric family

Value

a vector of n observations

Examples

n <- 100L
# the lower and upper bound are either scalar,
# or else vectors of length n
r_dtrunc_elife(n = n, scale = 1, shape = -0.1,
               lower = pmax(0, runif(n, -0.5, 1)),
               upper = runif(n, 6, 10),
               family = "gp")
#>   [1] 1.62121572 1.11804954 4.33255589 0.83835790 0.13231727 1.94260456
#>   [7] 2.27911296 1.94090852 1.19333773 0.65557517 2.41688603 1.65583718
#>  [13] 0.99436608 3.71770534 0.68246762 0.93225817 1.75530002 2.62241438
#>  [19] 2.98594635 0.98569869 3.26188894 0.17332777 1.01595047 0.59317293
#>  [25] 1.12461534 0.11450537 1.43716884 0.30784360 0.95167624 1.24080906
#>  [31] 1.11532548 0.91999018 1.24657894 1.24277033 0.64245335 1.05051160
#>  [37] 1.60745654 3.18504860 1.34932180 1.26858002 1.08367748 1.87407147
#>  [43] 0.91178626 0.92648922 0.41180966 0.25294943 0.03361708 0.38008247
#>  [49] 0.75943352 0.87160388 0.56163329 0.11528010 3.24938795 0.70004976
#>  [55] 0.12216051 1.69243139 0.59687959 1.96621880 1.90705234 2.17745745
#>  [61] 1.06121488 1.69239797 1.54952643 0.93228826 1.77547905 1.66744532
#>  [67] 1.34999894 0.67481785 0.49456312 1.50166854 1.97636936 0.63534397
#>  [73] 1.16347268 0.20915566 1.38482877 0.65478358 0.77076789 1.61928217
#>  [79] 0.25315413 3.63547594 0.88800467 1.43233377 0.93357379 0.43999700
#>  [85] 0.26157818 1.94381112 1.22233500 1.04506591 0.31968657 0.33210434
#>  [91] 2.91629476 2.70034110 2.80136086 1.42385645 1.20584851 0.11976767
#>  [97] 2.45313654 0.20350265 1.00438680 0.27937338