Sample observations from an interval truncated excess lifetime distribution

Sample observations from an interval truncated excess lifetime distribution

Usage

r_dtrunc_elife(
  n,
  scale,
  shape,
  lower,
  upper,
  family = c("exp", "gp", "gomp", "gompmake", "weibull", "extgp")
)

Arguments

n

sample size

scale

scale parameter(s)

shape

shape parameter(s)

lower

vector of lower bounds

upper

vector of upper bounds

family

string; choice of parametric family, either exponential (exp), Weibull (weibull), generalized Pareto (gp), Gompertz (gomp), Gompertz-Makeham (gompmake) or extended generalized Pareto (extgp).

Value

a vector of n observations

Examples

n <- 100L
# the lower and upper bound are either scalar,
# or else vectors of length n
r_dtrunc_elife(n = n, scale = 1, shape = -0.1,
               lower = pmax(0, runif(n, -0.5, 1)),
               upper = runif(n, 6, 10),
               family = "gp")
#>   [1] 2.49946336 2.01694956 2.73123586 0.28892848 0.67678126 1.69446274
#>   [7] 0.87512609 0.82394260 0.72849275 0.49896497 1.71371272 0.40255351
#>  [13] 3.85000965 0.95171903 1.15985232 1.04527785 1.62127355 0.85231925
#>  [19] 0.31549035 0.89895608 2.41055892 1.24156347 0.61155632 1.13099274
#>  [25] 1.46898319 1.50009070 1.22230854 0.32801049 3.20276150 0.36092852
#>  [31] 0.30528064 0.43120185 1.94962327 1.85248584 3.82907930 1.12011742
#>  [37] 0.01074014 0.45635476 1.24652132 2.33568908 1.67029532 0.91436935
#>  [43] 0.87434177 0.70263529 1.92628837 1.32049871 2.49933088 0.45158134
#>  [49] 1.44550803 1.96171239 0.27226772 1.53079531 1.47244980 0.74676488
#>  [55] 0.95316168 0.79622326 0.93637477 0.07333934 2.51355893 0.15077384
#>  [61] 1.08557492 0.29597259 2.17725111 1.28740885 0.26079413 2.63968047
#>  [67] 0.40034876 1.43253874 1.44326741 0.89539729 1.20056515 0.45265391
#>  [73] 0.92426801 0.12684600 0.59468883 0.80537332 1.00767616 1.46206219
#>  [79] 1.39808453 0.45974224 2.17056223 2.91409698 0.75939958 0.87506914
#>  [85] 1.23578673 1.81392181 0.86454096 2.20662364 2.20712921 0.92143902
#>  [91] 0.63322433 3.24692621 0.28253976 3.25267359 0.26484122 3.43992362
#>  [97] 0.84381955 2.55399575 0.29286344 0.96814956