Sample observations from a left-truncated right-censored excess lifetime distribution

Sample observations from a left-truncated right-censored excess lifetime distribution

Usage

r_ltrc_elife(
  n,
  scale,
  rate,
  shape,
  lower,
  upper,
  family = c("exp", "gp", "gomp", "gompmake", "weibull", "extgp", "gppiece",
    "extweibull", "perks", "beard", "perksmake", "beardmake")
)

Arguments

n

sample size

scale

scale parameter

shape

shape parameter(s)

lower

vector of lower bounds

upper

vector of upper bounds above which data are right-truncated

family

string; choice of parametric family, either exponential (exp), Weibull (weibull), generalized Pareto (gp), Gompertz (gomp) or extended generalized Pareto (extgp).

Value

a list with n observations dat and a logical vector of the same length with TRUE for right-censored observations and FALSE otherwise.

Examples

n <- 100L
# the lower and upper bound are either scalar,
# or else vectors of length n
r_ltrc_elife(n = n, scale = 5, shape = -0.1,
               lower = pmax(0, runif(n, -0.5, 1)),
               upper = 5,
               family = "gp")
#> $dat
#>   [1] 2.336863329 3.595901899 5.000000000 0.001810092 1.570244912 5.000000000
#>   [7] 5.000000000 5.000000000 5.000000000 0.039027097 5.000000000 1.734087296
#>  [13] 0.785274711 2.401133217 1.126958808 2.528427481 3.651339500 3.653942262
#>  [19] 0.534719465 4.550971549 5.000000000 2.220411329 0.677219374 2.799134667
#>  [25] 2.294262942 4.016549907 4.606724113 5.000000000 0.775277593 4.485614115
#>  [31] 5.000000000 3.814730486 1.787996884 1.217029423 4.619413740 5.000000000
#>  [37] 0.839067060 3.008559862 1.438535554 5.000000000 5.000000000 5.000000000
#>  [43] 3.052403886 5.000000000 5.000000000 5.000000000 0.270998576 1.249693290
#>  [49] 4.913267804 1.306190826 5.000000000 0.233899808 2.749061255 0.533078886
#>  [55] 5.000000000 5.000000000 2.139727644 3.144367906 5.000000000 5.000000000
#>  [61] 2.253813905 5.000000000 2.592846465 5.000000000 2.077304999 1.537841640
#>  [67] 5.000000000 1.730290696 5.000000000 3.927962050 1.660080752 2.769785325
#>  [73] 3.273827555 4.365875467 4.120968993 5.000000000 1.046266821 0.663331914
#>  [79] 5.000000000 3.856506834 0.363591396 1.094757570 5.000000000 5.000000000
#>  [85] 3.190864677 2.221471580 1.959137409 2.078210024 4.345968606 0.896874062
#>  [91] 0.659726767 1.890285003 2.784766471 5.000000000 1.191587161 4.918211529
#>  [97] 1.136610856 5.000000000 5.000000000 2.982274193
#> 
#> $rcens
#>   [1] FALSE FALSE  TRUE FALSE FALSE  TRUE  TRUE  TRUE  TRUE FALSE  TRUE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE  TRUE
#>  [37] FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE  TRUE FALSE FALSE
#>  [49] FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE  TRUE FALSE FALSE  TRUE  TRUE
#>  [61] FALSE  TRUE FALSE  TRUE FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE  TRUE  TRUE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE
#>  [97] FALSE  TRUE  TRUE FALSE
#>